% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperr8.R
\name{hyperr8_norm_run}
\alias{hyperr8_norm_run}
\title{Run all analyses using norm appoach}
\usage{
hyperr8_norm_run(
  all_data,
  nreps = 5,
  epsilon_lower = -Inf,
  nstep_dentist = 1000
)
}
\arguments{
\item{all_data}{A data frame with columns of time, rate, and perhaps citation, numerator, denominator, and/or total_time (other columns will be ignored).}

\item{nreps}{The number of times to randomize the data within each dataset.}

\item{epsilon_lower}{The lower bound for the epsilon parameter.}

\item{nstep_dentist}{The number of steps for the dentist algorithm.}
}
\value{
A data.frame of results with class hyperr8.
}
\description{
By default will run analyses on all datasets. It expects an input data frame with columns of time and rate, each row representing a single observation. To handle multiple datasets at once, have a citation column indicating the name for the dataset (which could be a full citation or as simple as "dataset1", "dataset2", etc.). You can also include a column for numerator (the number of events) and denominator (the number of opportunities for events). If you have a column for total_time, it will use that for randomization rather than time (this is important for things like phylogenetic trees, where the time over which events happen is not just the overall age of the tree but the sum of its branch lengths).
}
\details{
Data can be randomized to help suggest whether an observed pattern is spurious.
}
\examples{
library(hyperr8); car_data <- generate_car_simulation(); all_run <- hyperr8_norm_run(car_data)
subset(as.data.frame(all_run), datum_id==1)
library(ggplot2)
plot(all_run)
}
