% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperr8.R
\name{log.start}
\alias{log.start}
\title{Function to find a constant for which log(y+c) is as symmetric as possible at the alpha and 1-alpha percentiles}
\usage{
\method{log}{start}(y, alpha = 0.25, beta = 1, gamma = 0.01)
}
\arguments{
\item{y}{A vector of values}

\item{alpha}{The percentile to use for the lower bound}

\item{beta}{The percentile to use for extrapoloation}

\item{gamma}{The amount to extrapolate below the min}
}
\value{
A list with the offset and a code
}
\description{
This is done rather than a fixed log1p offset or a fixed tiny offset.
Code is from:
whuber (https://stats.stackexchange.com/users/919/whuber), Choosing c such that log(x + c) would remove skew from the population, URL (version: 2012-11-05): https://stats.stackexchange.com/q/41415
}
