% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperr8.R
\name{optimization_norm_over_all_data}
\alias{optimization_norm_over_all_data}
\title{Optimize rate model using noise model}
\usage{
optimization_norm_over_all_data(
  all_data,
  models = generate_all_models()[c(3, 5, 7), ],
  epsilon_lower = -Inf,
  nstep_dentist = 1000
)
}
\arguments{
\item{all_data}{A data frame with columns of time, rate, and citation.}

\item{models}{A data frame with columns of e, k, a, and description}

\item{epsilon_lower}{The lower bound for the epsilon parameter.}

\item{nstep_dentist}{The number of steps for the dentist algorithm.}
}
\value{
A list of results, one for each model and dataset.
}
\description{
This will dredge all the rate models for a given dataset. It expects column names of time, rate, and citation; optionally, it can also include a numerator, denominator, and/or total_time columns. It will return a list of results, one for each model and dataset.
}
\details{
By default it will use all models, but you can specify a subset of models to use.
}
