% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperr8.R
\name{optimization_over_all_data}
\alias{optimization_over_all_data}
\title{Optimize rate model}
\usage{
optimization_over_all_data(all_data, models = generate_all_models())
}
\arguments{
\item{all_data}{A data frame with columns of time, rate, and citation.}

\item{models}{A data frame with columns of e, k, a, and description}
}
\value{
A list of results, one for each model and dataset.
}
\description{
This will dredge all the rate models for a given dataset. It expects column names of time, rate, and citation; optionally, it can also include a numerator, denominator, and/or total_time columns. It will return a list of results, one for each model and dataset.
}
\details{
By default it will use all models, but you can specify a subset of models to use.
}
