% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperr8.R
\name{plot_proportion_with_offset}
\alias{plot_proportion_with_offset}
\title{Plot contributions of each component}
\usage{
plot_proportion_with_offset(x, scaling_factor = 0.3)
}
\arguments{
\item{x}{The output from summarize_all_fitted_models, filtered to a single dataset and model}

\item{scaling_factor}{The scaling factor for the plot, which will be used to determine the height of the ribbons.}
}
\value{
A ggplot2 object.
}
\description{
This will plot the contributions of each component.
}
