% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperr8.R
\name{print.hyperr8}
\alias{print.hyperr8}
\title{Print hyperr8 results

This will summarize the results from hyperr8_run.}
\usage{
\method{print}{hyperr8}(x, ...)
}
\arguments{
\item{x}{The output from hyperr8_run.}

\item{...}{Other arguments to pass to the printing function.}
}
\value{
A data.frame with summary information.
}
\description{
Print hyperr8 results

This will summarize the results from hyperr8_run.
}
