% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperr8.R
\name{summarize_all_fitted_models}
\alias{summarize_all_fitted_models}
\title{Generate summary information for all fitted models}
\usage{
summarize_all_fitted_models(minimization_approach_result)
}
\arguments{
\item{minimization_approach_result}{A list of results from optimization_over_all_data}
}
\value{
A data frame with summary information for all fitted models.
}
\description{
This will generate a data frame with summary information for all fitted models.
}
