% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperr8.R
\name{summary.hyperr8}
\alias{summary.hyperr8}
\title{Summarize hyperr8 results}
\usage{
\method{summary}{hyperr8}(object, ...)
}
\arguments{
\item{object}{The output from hyperr8_run.}

\item{...}{Other arguments to pass to the summary function.}
}
\value{
A list with summary information.
}
\description{
This will summarize the results from hyperr8_run.
}
