% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperr8.R
\docType{data}
\name{yule_sim}
\alias{yule_sim}
\title{This is data to be included in the package}
\format{
a data.frame with 25,000 rows and 8 columns

This has a dataset of estimated speciation rate for every tree simulated from a pure birth simulation with a speciation rate 0.1 and total time pulled from a distribution uniform on a log space ranging from 1 to 50 time units. Trees started with two taxa; trees with no speciation events ended with two taxa, having an inferred speciation rate of 0.
}
\usage{
data(yule_sim)
}
\description{
This is data to be included in the package
}
\keyword{data}
