% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperr8.R
\name{generate_car_simulation}
\alias{generate_car_simulation}
\title{Car simulation
This will generate a data frame with simulated car data. By default, it will have 1000 rows representing 1000 cars. Each car will have a simulated driving time and a simulated driving speed. The driving time and driving speed will be correlated. The correlation will be 0.99 by default.}
\usage{
generate_car_simulation(
  mean_driving_time = 3,
  mean_driving_speed = 70,
  sd_driving_time = 0.2,
  sd_driving_speed = 30,
  cor_driving = 0.99,
  n = 1000
)
}
\arguments{
\item{mean_driving_time}{The mean driving time for all cars.}

\item{mean_driving_speed}{The mean driving speed for all cars.}

\item{sd_driving_time}{The standard deviation of the driving time for the cars.}

\item{sd_driving_speed}{The standard deviation of the driving speed for the cars.}

\item{cor_driving}{The correlation between driving time and driving speed.}

\item{n}{The number of cars to simulate.}
}
\value{
A data frame with simulated car data.
}
\description{
Car simulation
This will generate a data frame with simulated car data. By default, it will have 1000 rows representing 1000 cars. Each car will have a simulated driving time and a simulated driving speed. The driving time and driving speed will be correlated. The correlation will be 0.99 by default.
}
\examples{
library(hyperr8)
car_data <- generate_car_simulation()
plot(car_data$time, car_data$distance)
plot(car_data$time, car_data$estimated_speed)
}
